+++ src/components/ActiveButton/ActiveButton.tsx
@@ -1,24 +1,6 @@
-import Theme from '@/styles/Theme';
-import styled from '@emotion/styled';
+import { ActiveButtonProps } from '@/types/components';
 
-//#TODO : font-size 변경하기
-interface ActiveButtonProps {
-  buttonText: string;
-  fontSize: 'largeTitle' | 'mediumTitle' | 'smallTitle' | 'smallContent';
-}
-
-function getFontSize(fontSize: ActiveButtonProps['fontSize']) {
-  return Theme.fontSize[fontSize];
-}
-
-const ButtonStyled = styled.button<{ fontSize: ActiveButtonProps['fontSize'] }>`
-  padding: 0.8rem 1rem;
-  border: 0;
-  border-radius: 0.5rem;
-  background-color: ${Theme.color.tSemiActive};
-  color: white;
-  font-size: ${(props) => getFontSize(props.fontSize)};
-`;
+import { ButtonStyled } from './ActiveButton.style';
 
 const ActiveButton = ({ buttonText, fontSize }: ActiveButtonProps) => {
   return <ButtonStyled fontSize={fontSize}>{buttonText}</ButtonStyled>;
