+++ src/routes/index.tsx
@@ -1,5 +1,4 @@
 import App from '@/App';
-
 import FormLayout from '@/pages/FormLayout/FormLayout';
 import Layout from '@/pages/Layout/Layout';
 import LoginPage from '@/pages/LoginPage/LoginPage';
@@ -10,7 +9,7 @@ import ProfilePage from '@/pages/ProfilePage';
 import RegisterPage from '@/pages/RegisterPage/RegisterPage';
 import ClubEventPage from '@/pages/club/ClubEventPage';
 import ClubHomePage from '@/pages/club/ClubHomePage';
-import CreateClubPage from '@/pages/club/CreateClubPage';
+import CreateClubPage from '@/pages/club/CreateClubPage/CreateClubPage';
 import ManageClubPage from '@/pages/club/ManageClubPage';
 import CheckFormPage from '@/pages/event/CheckFormPage';
 import ChoiceEventPage from '@/pages/event/ChoiceEventPage';
