+++ src/types/components.ts
@@ -1,0 +1,6 @@
+interface ActiveButtonProps {
+  buttonText: string;
+  fontSize: 'largeTitle' | 'mediumTitle' | 'smallTitle' | 'smallContent';
+}
+
+export { ActiveButtonProps };
