+++ src/utils/getFontSize.ts
@@ -1,0 +1,6 @@
+import Theme from '@/styles/Theme';
+import { ActiveButtonProps } from '@/types/components';
+
+export function getFontSize(fontSize: ActiveButtonProps['fontSize']) {
+  return Theme.fontSize[fontSize];
+}
